feat: 파일 저장 변경 및 중복 코드 리팩토링

+++ src/main/java/com/example/DCRW/dto/board/FileDto.java
@@ -13,5 +13,5 @@ public class FileDto {
     private String filePath;
     private String fileUrl;
     private String fileType;
-    private int fileSize;
+    private Long fileSize;
 }
